+++ package.json
@@ -157,7 +157,7 @@
     },
     "husky": {
         "hooks": {
-            "pre-commit": "run-p changelog schema schema-higlass schema-theme schema-template format",
+            "pre-commit": "run-p changelog schema schema-higlass schema-theme schema-template format && git add .",
             "commit-msg": "commitlint -E HUSKY_GIT_PARAMS"
         }
     }
